Added parent class for all support dolls
