+++ krie.py
@@ -2,9 +2,11 @@ import pygame
 import time
 import math
 import spritesheet
+from parent_classes.support_dolls import *
 
 
-class Krie(pygame.sprite.Sprite):
+
+class Krie(pygame.sprite.Sprite, Support):
     def __init__(self, game):
         super().__init__()
         self.game = game
@@ -21,60 +23,19 @@ class Krie(pygame.sprite.Sprite):
 
 
     def update(self,deltatime, player_action, player_x, player_y):
-        self.current_time += deltatime
+         self.update_movement(deltatime, player_action, player_x, player_y, self.animate)
 
-        # Check player direction
-        direction_x = player_action["right"] - player_action["left"]
-        direction_y = player_action["down"] - player_action["up"]
-
-        # Cooldown for attack
-        if self.current_time > 3:
-            self.attack = True
-            self.attack_cooldown += deltatime
-            if self.attack_cooldown > 0.8:
-                self.attack = False
-                self.attack_cooldown = 0
-                self.current_time = 0
-        
-        # Move towards player always
-        if not self.attack:
-            self.move(player_x, player_y)
-
-        self.animate(deltatime, direction_x, direction_y, self.step_distance)
     
     def render(self, display):
-        display.blit(self.image, (self.krie_vector.x, self.krie_vector.y))
+        display.blit(self.image, (self.doll_vector.x, self.doll_vector.y))
         # pygame.draw.rect(display, (255,255,255), self.rect,2)
 
 
     def animate(self, deltatime, direction_x, direction_y, distance):
         self.last_frame_update += deltatime
 
-        # Support doll idle
-        if not(direction_x or direction_y) and (self.attack == False):
-            if self.current_anim_list == self.right_sprites or self.current_anim_list == self.walk_right or self.current_anim_list == self.attack_right:
-                self.current_anim_list = self.right_sprites
-                self.image = self.current_anim_list[self.current_frame_unique]
-            elif self.current_anim_list == self.left_sprites or self.current_anim_list == self.walk_left or self.current_anim_list == self.attack_left:
-                self.current_anim_list = self.left_sprites
-                self.image = self.current_anim_list[self.current_frame_unique]
-            if self.last_frame_update > 0.5:
-                self.current_frame_unique = (self.current_frame_unique + 1) % len(self.right_sprites)
-                self.last_frame_update = 0 
-            return
-        
-        # Support doll walking
-        if direction_x and self.attack == False:
-            if direction_x > 0:
-                if distance > 0.4:
-                    self.current_anim_list = self.walk_right
-                else:
-                    self.current_anim_list = self.right_sprites
-            else: 
-                if distance > 0.4:
-                    self.current_anim_list = self.walk_left
-                else:
-                    self.current_anim_list =self.left_sprites
+        # Code for all support doll's walking and idle animation
+        self.idle_walking(direction_x, direction_y, distance)
 
         # walk animation after attacking
         if direction_y != 0 and (self.image == self.attack_right[self.current_frame]) and self.attack == False: 
@@ -85,13 +46,11 @@ class Krie(pygame.sprite.Sprite):
         # Support doll attacking animation
         if self.attack == True and (self.current_anim_list == self.right_sprites or self.current_anim_list == self.walk_right):
             self.current_frame = 0
-            self.current_anim_list.clear
             self.fps = 0.1
             self.current_anim_list = self.attack_right[0]
             self.current_anim_list = self.attack_right
         if self.attack == True and (self.current_anim_list == self.left_sprites or self.current_anim_list == self.walk_left):
             self.current_frame = 0
-            self.current_anim_list.clear
             self.fps = 0.1
             self.current_anim_list = self.attack_left[0]
             self.current_anim_list = self.attack_left
@@ -118,34 +77,7 @@ class Krie(pygame.sprite.Sprite):
                 self.image = self.current_anim_list[self.current_frame]
                 self.last_frame_update = 0 
 
-        
 
-
-    # This code is to make sure Support doll is always in range of Player
-    def move(self, player_x, player_y):
-
-        self.torres_vector = pygame.math.Vector2(player_x, player_y)
-        self.krie_vector = pygame.math.Vector2(self.rect.x, self.rect.y)
-        self.step_distance = 0
-        self.min_distance = 400
-        self.max_distance = 500
-        # print(self.follower_vector)
-
-
-        # distance_to returns the pythagorean distance between two points
-        self.distance = self.krie_vector.distance_to(self.torres_vector)
-        
-        if self.distance > self.min_distance:
-            self.direction_vector = (self.torres_vector - self.krie_vector) / self.distance
-            self.min_step        = max(0, self.distance - self.max_distance)
-            self.max_step        = self.distance - self.min_distance
-            #step_distance       = min(max_step, max(min_step, VELOCITY))
-            self.step_distance   = self.min_step + (self.max_step - self.min_step) 
-            self.krie_vector += self.direction_vector * self.step_distance * 0.1
-            self.rect.x, self.rect.y = self.krie_vector.x, self.krie_vector.y
-
-
-        
 
     def load_sprites(self):
         self.right_sprites, self.left_sprites = [], []
